# src/server/coct_sr_info_server/models/image_recognition_schema.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class ImageRecognitionSchema(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ImageRecognitionSchema - a model defined in OpenAPI

        type: The type of this ImageRecognitionSchema.
        subtype: The subtype of this ImageRecognitionSchema.
        description: The description of this ImageRecognitionSchema [Optional].
        url: The url of this ImageRecognitionSchema [Optional].
        confidence: The confidence of this ImageRecognitionSchema.
    """

    type: str = Field(alias="type")
    subtype: str = Field(alias="subtype")
    description: Optional[str] = Field(alias="description", default=None)
    url: Optional[str] = Field(alias="url", default=None)
    confidence: float = Field(alias="confidence")

ImageRecognitionSchema.update_forward_refs()