# src/server/coct_sr_info_server/models/service_request_schema.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class ServiceRequestSchema(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ServiceRequestSchema - a model defined in OpenAPI

        id: The id of this ServiceRequestSchema.
        sr_status: The sr_status of this ServiceRequestSchema.
        message: The message of this ServiceRequestSchema.
        type: The type of this ServiceRequestSchema.
        subtype: The subtype of this ServiceRequestSchema.
        suburb: The suburb of this ServiceRequestSchema [Optional].
        latitude: The latitude of this ServiceRequestSchema [Optional].
        longitude: The longitude of this ServiceRequestSchema [Optional].
    """

    id: str = Field(alias="id")
    sr_status: str = Field(alias="sr_status")
    message: str = Field(alias="message")
    type: str = Field(alias="type")
    subtype: str = Field(alias="subtype")
    suburb: Optional[str] = Field(alias="suburb", default=None)
    latitude: Optional[float] = Field(alias="latitude", default=None)
    longitude: Optional[float] = Field(alias="longitude", default=None)

ServiceRequestSchema.update_forward_refs()