# src/server/coct_sr_info_server/models/service_request_stats_schema.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from coct_sr_info_server.models.service_request_stats_schema_area_stats import ServiceRequestStatsSchemaAreaStats


class ServiceRequestStatsSchema(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ServiceRequestStatsSchema - a model defined in OpenAPI

        name: The name of this ServiceRequestStatsSchema [Optional].
        area_stats: The area_stats of this ServiceRequestStatsSchema [Optional].
    """

    name: Optional[str] = Field(alias="name", default=None)
    area_stats: Optional[ServiceRequestStatsSchemaAreaStats] = Field(alias="area_stats", default=None)

ServiceRequestStatsSchema.update_forward_refs()